astoria-ai/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
openai>=1.3.5
httpx>=0.24.0
jinja2>=3.1.2
PyMuPDF>=1.21.1
chromadb>=0.4.14
passlib[bcrypt]>=1.7.4
sqlite3
requests>=2.31.0
uuid
typing_extensions>=4.5.0
loguru>=0.7.0
